Switch screenshot mechanism to using libvirt screenshots.
This gets rid of an annoying external dependency, and
makes for some simpler code.  Unfortunately I was not able
to get it to work on a Fedora-14 host with libvirt 0.9.6,
but I have not investigated the causes.

Signed-off-by: Chris Lalancette <clalance@redhat.com>
(cherry picked from commit a0cd2832f4ddb90878272f3b283a292edc95507b)

Conflicts:
	oz.spec.in