add VIR_ERR_OPERATION_INVALID as valid complete install error [BZ 755830]
In F16 (and possibly earlier) this is one of the errors that libvirt
throws when you attempt to operate on a domain that has shut down.
This needs to be one of the "acceptable errors" we catch as an indication
that the JEOS install has finished.
(cherry picked from commit d4e90b19d13c5c52a860d0783facd5ee1c5de873)